backend/requirements.txt
# Core Django
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.6.0
python-decouple==3.8

# PDF Processing
PyPDF2==3.0.1
PyMuPDF==1.24.0
pdf2image==1.17.0
pillow==12.0.0

# Google Gemini AI (updated for langchain compatibility)
google-generativeai>=0.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5

# HTTP Requests (for Eleven Labs API)
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11

# Utilities
httplib2==0.31.0
cachetools==6.2.1
uritemplate==4.2.0
pyparsing==3.2.5
rsa==4.9.1
sqlparse==0.5.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# LangChain and Environment
python-dotenv==1.0.0
langchain==0.3.27
langchain-google-genai==2.0.10
langchain-core>=0.3.15

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
# ============================================================================
# EMAIL AUTOMATION DEPENDENCIES
# ============================================================================

# --- Google Gemini AI (Required) ---
# Official Google Generative AI SDK for Gemini models
google-genai>=0.2.0

# --- Environment & Configuration ---
# Securely load API keys and credentials from .env file
python-dotenv>=1.0.0

# ============================================================================
# OPTIONAL: LangChain Integration (for ai_agent.py)
# ============================================================================
# Uncomment these if you're also using LangChain in other scripts

# LangChain core library for agents and chains
langchain>=0.1.0

# LangChain integration with Google Gemini
langchain-google-genai>=2.0.0

# Pydantic for data validation
pydantic>=2.0.0

# Production dependencies
whitenoise==6.6.0
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0